+++ docs/Changelog.md
@@ -105,6 +105,7 @@ When test preprocessing is enabled, Ceedling discovers whether your toolchain su
    - #1115 Fixed Gcov plugin for Gcovr configuration file handling.
    - #1161 Fixed “Found no coverage results” when system headers are processed by coverage.
 - #1144 Fixed flags for release build linking.
+- #1169 Fixed linking with :libraries ↳ :release.
 
 ## ⚠️ Changed